Index replacements by string, not `Annotation`
This makes our data structures a tad more reusable. This commit
should hopefully also improve the performance a little, mainly by
replacing `m` concatenations of strings of lengths in the order
of `d`, where `m` is the number of mentions and `d` length of
the document, by an order of `m` substring operations followed
by one `str.join`.